requirements.txt
asgiref==3.11.1
Django==6.0.4
djangorestframework==3.17.1
django-cors-headers==4.7.0
numpy==2.4.4
opencv-python==4.13.0.92
sqlparse==0.5.5
tzdata==2026.1
Pillow==11.2.1
requests==2.32.3
PyJWT==2.10.1
